requirements.txt
Django==1.11.5
django-filter==1.0.4
django-rest-framework==0.1.0
djangorestframework==3.7.1
pkg-resources==0.0.0
psycopg2==2.7.3.1
pytz==2017.2
